third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/scoped-custom-element-registry-customelementregistry-attribute-expected.txt

This is a testharness.js-based test.
[FAIL] HTML parser should create a builtin element with null registry if customelementregistry is set
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] document.createElement should create a builtin element with null registry if customElement is set to null
  Failed to execute 'createElement' on 'Document': Failed to read the 'customElementRegistry' property from 'ElementCreationOptions': Failed to convert value to 'CustomElementRegistry'.
[FAIL] attchShadow on a builtin element with null customElementRegistry should create a ShadowRoot with null registry
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Cloning a builtin element with null regsitry should create an element with null registry
  Failed to execute 'createElement' on 'Document': Failed to read the 'customElementRegistry' property from 'ElementCreationOptions': Failed to convert value to 'CustomElementRegistry'.
[FAIL] HTML parser should create a custom element candidate with null registry if customelementregistry is set
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] document.createElement should create a custom element candidate with null registry if customElement is set to null
  Failed to execute 'createElement' on 'Document': Failed to read the 'customElementRegistry' property from 'ElementCreationOptions': Failed to convert value to 'CustomElementRegistry'.
[FAIL] attchShadow on a custom element candidate with null customElementRegistry should create a ShadowRoot with null registry
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Cloning a custom element candidate with null regsitry should create an element with null registry
  Failed to execute 'createElement' on 'Document': Failed to read the 'customElementRegistry' property from 'ElementCreationOptions': Failed to convert value to 'CustomElementRegistry'.
[FAIL] HTML parser should create a custom element with null registry if customelementregistry is set
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] document.createElement should create a custom element with null registry if customElement is set to null
  Failed to execute 'createElement' on 'Document': Failed to read the 'customElementRegistry' property from 'ElementCreationOptions': Failed to convert value to 'CustomElementRegistry'.
[FAIL] attchShadow on a custom element with null customElementRegistry should create a ShadowRoot with null registry
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Cloning a custom element with null regsitry should create an element with null registry
  Failed to execute 'createElement' on 'Document': Failed to read the 'customElementRegistry' property from 'ElementCreationOptions': Failed to convert value to 'CustomElementRegistry'.
Harness: the test ran to completion.
